add fetch_iso script prototype
fetch_iso - new script to manage ISO image repository mirrors.
            by this scrips and keepname=0 args, customers can receive in advance
            images of all (or individual) profiles in ISOs directories.
            This can be useful when you want to insure yourself and have a
            mirror of all the necessary ISO on the host.
            by this script and keepname=1 args, customers can receive in advance
            images of all (or individual) profiles in ORIGINAL names -
            as they are named on the developers/official website. This can be
            useful when you want build local/regional MIRRORS where can go
            other CBSD/hosters.

In connection with this, we separate iso_site variables in profiles - as
 official site, and cbsd_iso_mirrors variables in profiles - as official
CBSD mirrors. This is necessary for the information from which sources we can get
md5sum for ISO. Since $cbsd_iso_mirrors will contain third-party mirrors,
еhey will not participate in the generation of MD5 sum.

This work is necessary, since some users expressed their desire to help
the project and add several regional mirrors for the needs of CBSD.

work-in-progress..