jail: populate_cdir,switch_baserw: always use realpath
In some cases, the source and target directories can be in the form of symbolic links.
In these cases, symbolic links are not handled correctly by cpio with follow notices:
  cpio: <src>: Cannot extract through symlink <dst>
We must make sure that we always go to the actual real directory before cpio from curdir,
therefore we process src/dst directory by realpath(1).
  Issue #287, Reported by: Juronski. Thanks!