nc.subr: change params=value parsing method
old behaviour via ARG="${1%%=*}" and VAL="${1##*=}" is not suitable for values
with "=" symbol, e.g:  params="value=2" get 2 for VAL, not value=2.
up to this point we were lucky that there were no such parameters, but
everything changed with the @am11 arrival ;-) who found the situation where
"=" character may be present in value, Issue #367.
So its time to make this method more reliable