return CBSD version as cbsdver variable into workdir
after adding support for multiple workdir,
we have deleted many obsolete files in workdir.
in one of these files (cbsd.conf) we kept the version
with which this environment was initialized.
Old version number is used by migration/initenv so
we have to return it.
Now this variable is stored in local SQLite3 and exported
into ~cbsd/nc.inventory file

Issue #404