move mod_redis data to personal module dir
In this way we can enable mod_redis by module.conf as other modules:
  echo "cbsd_redis.d" >> ~cbsd/etc/modules.conf

in addition, it allows subsequently not to install
the module data, if this is not necessary.

probably later we can introduce command to work with modules, e.g:

  cbsd module install vncterm
  cbsd module upgrade cbsd_redis
  ..

which use archives from the public repository, e.g. GITHUB.