natcfg should give three options
natcfg allow CBSD users to configure internal simple NAT function.
on launch, the script asks if the user wants to configure nat.
If the nat was previously turned on and the user wants to turn it off, he must answer "yes"
to go to nat engine selection (or disable).
In this case, the answer "no" is used as "cancel"/"exit"
and this can be confusing as this can be interpreted as "no -> disable NAT".

Add "cancel" options for natcfg and disable, when users choose "no".

Besides, add no interactive flags to rc.d/cbsdd to rule out any
question on boot stage.

Pointed out: kr0m, Issue #521