dhcpd: addr/32 list support, lock/lease time
valid ip4pool: IP/prefix, x.y.i.{a-b}, ip1 ip2 ip3
add lock/lease time to avoid collisions when running
the script in parallel and at the same time.

we block the IP through the file for 30 seconds.
this is especially important for CBSDfile and multiple
env with ip4_addr="DHCP"