do not manage ~cbsd/.ssh/authorized_keys due to security reason
historically, cbsd user was only used for remote commands after "cbsd node mode=add" operations.
At the moment there are several frontend systems for CBSD that require adding an unprivileged user to the cbsd group.
This can lead to leakage of the private/public key through the frontend and access to the server
if the server administrator has not closed the SSH and does not block the "cbsd" user of the through SSH.
Accordingly, if administrator want "cbsd node mode=add", he must copy the public key into ~cbsd/.ssh/authorized_keys.