Correct some spelling, grammar & mechanical stuff (#803)
* Harmonize conjugation and capitalization in help

Descriptions in `cbsd help` are now capitalized, except for those
starting with a proper noun that isn’t usually capitalized.
Descriptions that indicate action use a bare infinitive;
everything else was left as-is.

* Change “Use menu for” to “Use menu to”

If the verbs were nominalized, “for” would’ve been fine.
Also, some “and”s were missing in the same prompts.